+++ src/monaco-loader.ts
@@ -1,8 +1,11 @@
-//export * from "@hediet/monaco-editor-react/dist/monaco-loader";
+import type * as monaco from "monaco-editor";
 
-import * as monaco from "monaco-editor";
 export type Monaco = typeof monaco;
 
 export async function loadMonaco(): Promise<Monaco> {
-	return monaco;
+	if (process.env.USE_CDN_FOR_MONACO === "1") {
+		return require("@hediet/monaco-editor-react/dist/monaco-loader").loadMonaco();
+	} else {
+		return require("monaco-editor");
+	}
 }